tests/api-resources/accounts/gateway/gateway.test.ts
// File generated from our OpenAPI spec by Stainless.

import Cloudflare from 'cloudflare';
import { Response } from 'node-fetch';

const cloudflare = new Cloudflare({
  apiEmail: 'dev@cloudflare.com',
  apiKey: 'my-cloudflare-api-key',
  apiToken: 'my-cloudflare-api-token',
  userServiceKey: 'my-cloudflare-user-service-key',
  baseURL: process.env['TEST_API_BASE_URL'] ?? 'http://127.0.0.1:4010',
});

describe('resource gateway', () => {
  // skipped: tests are disabled for the time being
  test.skip('zeroTrustAccountsCreateZeroTrustAccount: only required params', async () => {
    const responsePromise = cloudflare.accounts.gateway.zeroTrustAccountsCreateZeroTrustAccount(
      '699d98642c564d2e855e9661899b7252',
      { account_id: '699d98642c564d2e855e9661899b7252' },
    );
    const rawResponse = await responsePromise.asResponse();
    expect(rawResponse).toBeInstanceOf(Response);
    const response = await responsePromise;
    expect(response).not.toBeInstanceOf(Response);
    const dataAndResponse = await responsePromise.withResponse();
    expect(dataAndResponse.data).toBe(response);
    expect(dataAndResponse.response).toBe(rawResponse);
  });

  // skipped: tests are disabled for the time being
  test.skip('zeroTrustAccountsCreateZeroTrustAccount: required and optional params', async () => {
    const response = await cloudflare.accounts.gateway.zeroTrustAccountsCreateZeroTrustAccount(
      '699d98642c564d2e855e9661899b7252',
      { account_id: '699d98642c564d2e855e9661899b7252' },
    );
  });

  // skipped: tests are disabled for the time being
  test.skip('zeroTrustAccountsGetZeroTrustAccountInformation', async () => {
    const responsePromise = cloudflare.accounts.gateway.zeroTrustAccountsGetZeroTrustAccountInformation(
      '699d98642c564d2e855e9661899b7252',
    );
    const rawResponse = await responsePromise.asResponse();
    expect(rawResponse).toBeInstanceOf(Response);
    const response = await responsePromise;
    expect(response).not.toBeInstanceOf(Response);
    const dataAndResponse = await responsePromise.withResponse();
    expect(dataAndResponse.data).toBe(response);
    expect(dataAndResponse.response).toBe(rawResponse);
  });

  // skipped: tests are disabled for the time being
  test.skip('zeroTrustAccountsGetZeroTrustAccountInformation: request options instead of params are passed correctly', async () => {
    // ensure the request options are being passed correctly by passing an invalid HTTP method in order to cause an error
    await expect(
      cloudflare.accounts.gateway.zeroTrustAccountsGetZeroTrustAccountInformation(
        '699d98642c564d2e855e9661899b7252',
        { path: '/_stainless_unknown_path' },
      ),
    ).rejects.toThrow(Cloudflare.NotFoundError);
  });
});
